revision

+++ database/migrations/2023_03_29_142957_create_product_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('products', function (Blueprint $table) {
             $table->id();
-            $table->integer('vendorId');
+            $table->bigInteger('vendorId');
             $table->string('imagePath');
             $table->string('productName', 100);
             $table->string('productPrice');
